renamed files, adding pageIndex and new sounds

+++ utils.cpp
@@ -20,7 +20,7 @@ bool Utils::collision(QLabel* element1, QLabel* element2)
     return collisionDetected;
 }
 
-void Utils::delay(float secondsToWait)
+void Utils::delay(double secondsToWait)
 {
         QEventLoop loop;
         QTimer t;
